Got a silly animation for the ground impact window working

+++ simulation.h
@@ -22,8 +22,8 @@ public:
     Simulation();
     ~Simulation();
     void setGravity(float, float);
-    void createMob(string, int, int, int, int);
-    void createMob(string, int, int, int, int, string);
+    void createMob(string, float, float, float, float);
+    void createMob(string, float, float, float, float, string);
     QImage step();
     void createExplosion(Mob*);
 };
